// generated_new/post/model/post-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UserVO } from './user-vo';

export interface PostVO {
    'id'?: string;
    'title'?: string;
    'content'?: string;
    'thumbNum'?: number;
    'favourNum'?: number;
    'userId'?: string;
    'createTime'?: string;
    'updateTime'?: string;
    'tagList'?: Array<string>;
    'user'?: UserVO;
    'hasThumb'?: boolean;
    'hasFavour'?: boolean;
}