// generated_new/question/model/question-set-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { QuestionVO } from './question-vo';
// May contain unused imports in some cases
// @ts-ignore
import type { UserVO } from './user-vo';

export interface QuestionSetVO {
    'id'?: string;
    'title'?: string;
    'description'?: string;
    'tags'?: Array<string>;
    'questionNum'?: number;
    'userId'?: string;
    'favourNum'?: number;
    'createTime'?: string;
    'userVO'?: UserVO;
    'questions'?: Array<QuestionVO>;
}